adjust for new url

+++ src/main/java/com/fiap/streamingvideo/model/VideoStatisticsDTO.java
@@ -1,9 +1,9 @@
 package com.fiap.streamingvideo.model;
 
 public record VideoStatisticsDTO(
-    long totalVideos,
-    long totalFavorited,
-    double averageViews
+    Integer totalVideos,
+    Integer totalFavorited,
+    Double averageViews
 ) {
 
 }
